Took out code with specifications from AnimalServiceImpl
