+++ src/main/java/org/example/animalchipization/service/animal/AnimalValidator.java
@@ -5,9 +5,11 @@ import org.example.animalchipization.entity.Animal;
 import org.example.animalchipization.entity.AnimalType;
 import org.example.animalchipization.entity.Location;
 import org.example.animalchipization.service.Validator;
+import org.springframework.data.jpa.domain.Specification;
 
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Stream;
 
 /**
  * @author Aleksey
@@ -29,4 +31,7 @@ public interface AnimalValidator extends Validator<Animal> {
     void checkVisitedLocationsEmpty(Long animalId);
 
     void setDeathTimeIfDead(Animal animal);
+
+    Stream<Animal> getSearchedAnimalStream(Specification<Animal> spec,
+                                           int limit, int offset);
 }
