+++ src/main/java/org/example/animalchipization/service/animal/impl/AnimalServiceImpl.java
@@ -1,16 +1,14 @@
 package org.example.animalchipization.service.animal.impl;
 
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.criteria.*;
 import lombok.RequiredArgsConstructor;
 import org.example.animalchipization.dto.animal.*;
 import org.example.animalchipization.entity.*;
 import org.example.animalchipization.mapper.animal.AnimalMapper;
 import org.example.animalchipization.repository.*;
-import org.example.animalchipization.repository.JpaSpecificationBuilder;
 import org.example.animalchipization.service.animal.AnimalService;
 import org.example.animalchipization.service.animal.AnimalValidator;
 import org.example.animalchipization.service.location.LocationValidator;
+import org.example.animalchipization.repository.SpecificationFactory;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Stream;
 
 /**
  * @author Aleksey
@@ -30,7 +29,6 @@ public class AnimalServiceImpl implements AnimalService {
     private final AnimalMapper animalMapper;
     private final AnimalValidator animalValidator;
     private final LocationValidator locationValidator;
-    private final EntityManager entityManager;
 
 
     @Override
@@ -93,53 +91,13 @@ public class AnimalServiceImpl implements AnimalService {
 
     @Override
     @Transactional
-    public List<AnimalDtoOut> searchAnimals(AnimalSearchCriteria animalSearchCriteria, int limit, int offset) {
+    public List<AnimalDtoOut> searchAnimals(AnimalSearchCriteria animalSearchCriteria,
+                                            int limit, int offset) {
 
-        Specification<Animal> spec = Specification.where(
-                JpaSpecificationBuilder.<Animal>dateTimeFrom(
-                        Animal_.CHIPPING_DATE_TIME,
-                        animalSearchCriteria.startDateTime())
-        ).and(
-                JpaSpecificationBuilder.dateTimeTo(
-                        Animal_.CHIPPING_DATE_TIME,
-                        animalSearchCriteria.endDateTime())
-        ).and(
-                JpaSpecificationBuilder.join(
-                        Animal_.CHIPPER_ID,
-                        Account_.ACCOUNT_ID,
-                        animalSearchCriteria.chipperId(),
-                        JoinType.INNER)
-        ).and(
-                JpaSpecificationBuilder.join(
-                        Animal_.CHIPPING_LOCATION_ID,
-                        Location_.LOCATION_ID,
-                        animalSearchCriteria.chippingLocationId(),
-                        JoinType.INNER)
-        ).and(
-                JpaSpecificationBuilder.equal(
-                        Animal_.LIFE_STATUS,
-                        animalSearchCriteria.lifeStatus())
-        ).and(
-                JpaSpecificationBuilder.equal(
-                        Animal_.GENDER,
-                        animalSearchCriteria.gender())
-        );
+        Specification<Animal> spec =
+                SpecificationFactory.buildAnimalSearchSpec(animalSearchCriteria);
 
-        var cb = entityManager.getCriteriaBuilder();
-
-        var criteriaQuery = cb.createQuery(Animal.class);
-        var root = criteriaQuery.from(Animal.class);
-        var predicate = spec.toPredicate(root, criteriaQuery, cb);
-
-        if (predicate == null) criteriaQuery.select(root);
-        else criteriaQuery.where(predicate);
-        criteriaQuery.orderBy(cb.asc(root.get(Animal_.ANIMAL_ID)));
-
-        var typedQuery = entityManager.createQuery(criteriaQuery)
-                .setFirstResult(offset)
-                .setMaxResults(limit);
-
-        var animals = typedQuery.getResultStream();
+        Stream<Animal> animals = animalValidator.getSearchedAnimalStream(spec, limit, offset);
 
         return animals.map(animalMapper::toDto).toList();
     }
