+++ src/main/java/org/example/animalchipization/service/animal/impl/AnimalValidatorImpl.java
@@ -1,18 +1,17 @@
 package org.example.animalchipization.service.animal.impl;
 
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import static org.apache.commons.lang3.BooleanUtils.*;
 import org.example.animalchipization.dto.animal.AnimalDtoUpdate;
-import org.example.animalchipization.entity.Animal;
-import org.example.animalchipization.entity.AnimalType;
-import org.example.animalchipization.entity.Location;
-import org.example.animalchipization.entity.VisitedLocation;
+import org.example.animalchipization.entity.*;
 import org.example.animalchipization.enums.AnimalLifeStatus;
 import org.example.animalchipization.enums.error.BadRequestError;
 import org.example.animalchipization.enums.error.NotFoundError;
 import org.example.animalchipization.exception.RequestException;
 import org.example.animalchipization.repository.*;
 import org.example.animalchipization.service.animal.AnimalValidator;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
 
 import java.time.Instant;
@@ -20,6 +19,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Stream;
 
 /**
  * @author Aleksey
@@ -32,6 +32,7 @@ public class AnimalValidatorImpl implements AnimalValidator {
     private final AnimalTypeRepository animalTypeRepository;
     private final AccountRepository accountRepository;
     private final LocationRepository locationRepository;
+    private final EntityManager entityManager;
 
 
     @Override
@@ -126,5 +127,23 @@ public class AnimalValidatorImpl implements AnimalValidator {
         }
     }
 
+    public Stream<Animal> getSearchedAnimalStream(Specification<Animal> spec,
+                                                  int limit, int offset) {
 
+        var cb = entityManager.getCriteriaBuilder();
+
+        var criteriaQuery = cb.createQuery(Animal.class);
+        var root = criteriaQuery.from(Animal.class);
+        var predicate = spec.toPredicate(root, criteriaQuery, cb);
+
+        if (predicate == null) criteriaQuery.select(root);
+        else criteriaQuery.where(predicate);
+        criteriaQuery.orderBy(cb.asc(root.get(Animal_.ANIMAL_ID)));
+
+        var typedQuery = entityManager.createQuery(criteriaQuery)
+                .setFirstResult(offset)
+                .setMaxResults(limit);
+
+        return typedQuery.getResultStream();
+    }
 }
