+++ src/main/java/org/example/animalchipization/service/visitedLocation/impl/VisitedLocationServiceImpl.java
@@ -1,6 +1,5 @@
 package org.example.animalchipization.service.visitedLocation.impl;
 
-import jakarta.persistence.criteria.JoinType;
 import lombok.RequiredArgsConstructor;
 import org.example.animalchipization.dto.location.VisitedLocationSearchCriteria;
 import org.example.animalchipization.dto.visitedLocation.UpdateVisitedLocationDto;
@@ -11,8 +10,8 @@ import org.example.animalchipization.repository.VisitedLocationRepository;
 import org.example.animalchipization.service.animal.AnimalValidator;
 import org.example.animalchipization.service.location.LocationValidator;
 import org.example.animalchipization.service.visitedLocation.VisitedLocationService;
-import org.example.animalchipization.repository.JpaSpecificationBuilder;
 import org.example.animalchipization.service.visitedLocation.VisitedLocationValidator;
+import org.example.animalchipization.repository.SpecificationFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -39,24 +38,11 @@ public class VisitedLocationServiceImpl implements VisitedLocationService {
     @Override
     @Transactional
     public List<VisitedLocationDtoOut> searchLocations(
-            VisitedLocationSearchCriteria VisitedLocationSearchCriteria,
+            VisitedLocationSearchCriteria visitedLocationSearchCriteria,
             Pageable pageable) {
 
-        Specification<VisitedLocation> spec = Specification.where(
-                JpaSpecificationBuilder.<VisitedLocation>join(
-                        VisitedLocation_.ANIMAL,
-                        Animal_.ANIMAL_ID,
-                        VisitedLocationSearchCriteria.animalId(),
-                        JoinType.INNER)
-        ).and(
-                JpaSpecificationBuilder.dateTimeFrom(
-                        VisitedLocation_.DATE_TIME_OF_VISIT_LOCATION_POINT,
-                        VisitedLocationSearchCriteria.startDateTime())
-        ).and(
-                JpaSpecificationBuilder.dateTimeTo(
-                        VisitedLocation_.DATE_TIME_OF_VISIT_LOCATION_POINT,
-                        VisitedLocationSearchCriteria.endDateTime())
-        );
+        Specification<VisitedLocation> spec =
+                SpecificationFactory.buildVisitedLocationSearchSpec(visitedLocationSearchCriteria);
 
         Page<VisitedLocation> visitedLocationPage = visitedLocationRepository.findAll(spec, pageable);
